Modified per code review.
