+++ src/applications/vaos/appointment-list/components/RequestedAppointmentDetailsPage.jsx
@@ -31,7 +31,7 @@ function parseFakeFHIRId(id) {
 }
 
 function RequestedAppointmentDetailsPage({
-  appointmentDetails,
+  appointment,
   appointmentDetailsStatus,
   facilityData,
   fetchRequestDetails,
@@ -40,7 +40,6 @@ function RequestedAppointmentDetailsPage({
 }) {
   const { id } = useParams();
   const history = useHistory();
-  const appointment = appointmentDetails?.[id];
 
   useEffect(() => {
     if (pendingStatus !== FETCH_STATUS.succeeded) {
@@ -124,7 +123,6 @@ function RequestedAppointmentDetailsPage({
           <div>{appointment.reason}</div>
         </>
       )}
-
       {!isExpressCare && (
         <>
           <h2 className="vads-u-margin-top--2 vaos-appts__block-label">
@@ -133,7 +131,6 @@ function RequestedAppointmentDetailsPage({
           <div>{message}</div>
         </>
       )}
-
       <h2 className="vads-u-margin-top--2 vads-u-margin-bottom--0 vaos-appts__block-label">
         Your contact details
       </h2>
@@ -156,7 +153,6 @@ function RequestedAppointmentDetailsPage({
     </div>
   );
 }
-
 function mapStateToProps(state) {
   const {
     currentAppointment,
@@ -164,7 +160,6 @@ function mapStateToProps(state) {
     facilityData,
     pendingStatus,
   } = state.appointments;
-
   return {
     appointment: currentAppointment,
     appointmentDetailsStatus,
@@ -173,11 +168,9 @@ function mapStateToProps(state) {
     pendingStatus,
   };
 }
-
 const mapDispatchToProps = {
   fetchRequestDetails: actions.fetchRequestDetails,
 };
-
 export default connect(
   mapStateToProps,
   mapDispatchToProps,
